fix:コンポーネント分割 条件検索と、検索結果を分割
