+++ src/components/model/user/Condition/index.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import { Condition as ConditionImpl } from "./Condition";
+
+export const Condition = () => {
+  return <ConditionImpl />;
+};
