+++ src/components/model/user/List/ListHook.ts
@@ -1,42 +1,54 @@
-import { useForm, Controller } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { z } from "zod";
-import { useListSearchConditionMutators } from "@/components/store/useListSearchConditionState";
+import { useEffect, useState } from "react";
+import { UserListType, useUserList } from "@/components/usecase/useUserList";
+import { useListSearchConditionState } from "@/components/store/useListSearchConditionState";
 
-// Zod スキーマ
-const schema = z.object({
-  userId: z.string().optional(),
-  userName: z.string().optional(),
-});
-
-type FormData = z.infer<typeof schema>;
+// サンプルグリッドデータ
+const sampleData = [
+  {
+    id: 1,
+    userId: "system@co.jp",
+    userName: "システム管理者",
+    birthday: "2000/01/01",
+    age: 21,
+    gender: "男性",
+  },
+  {
+    id: 2,
+    userId: "user@co.jp",
+    userName: "ユーザー1",
+    birthday: "2000/01/01",
+    age: 21,
+    gender: "女性",
+  },
+];
 
 export const useListHook = () => {
-  // 検索条件のセッター
-  const { setListSearchCondition } = useListSearchConditionMutators();
+  const [rowData, setRowData] = useState<UserListType[]>([]);
 
-  // 検索ボタン押下アクション
-  const onValid = (form: FormData) => {
-    console.log(form);
+  // 検索条件
+  const condition = useListSearchConditionState();
 
-    const conditions: { key: string; value: string }[] = [];
-    if (form.userId) conditions.push({ key: "userId", value: form.userId });
-    if (form.userName)
-      conditions.push({ key: "userName", value: form.userName });
+  // データ取得処理
+  const { userListData, hasError, isLoading } = useUserList(condition);
 
-    const cond =
-      conditions.length === 0
-        ? ""
-        : "?" + conditions.map((item) => `${item.key}=${item.value}`).join("&");
+  // グリッドのカラム定義
+  const columns = [
+    { field: "userId", headerName: "ユーザーID", width: 150 },
+    { field: "userName", headerName: "ユーザー名", width: 150 },
+    { field: "birthday", headerName: "誕生日", width: 130 },
+    { field: "age", headerName: "年齢", width: 100 },
+    { field: "gender", headerName: "性", width: 100 },
+  ];
 
-    console.log("submit.condition", cond);
-    setListSearchCondition(cond);
-  };
+  //TODO 暫定処理 ID列の修正
+  useEffect(() => {
+    if (!userListData || !userListData.data) return;
+    const modData = userListData.data.map((item, i) => {
+      item.id = i + 1;
+      return item;
+    });
+    setRowData(modData);
+  }, [userListData]);
 
-  /** フォーム定義 */
-  const { handleSubmit, control } = useForm<FormData>({
-    resolver: zodResolver(schema),
-  });
-
-  return { handleSubmit, onValid, control };
+  return { rowData, hasError, isLoading, columns };
 };
