+++ src/components/page/user/ListPage.tsx
@@ -1,11 +1,32 @@
+import {
+  Paper,
+  TextField,
+  Box,
+  Button,
+  Typography,
+  Divider,
+  Container,
+} from "@mui/material";
 import { BaseLayout } from "@/components/model/layout";
-import React from "react";
 import { List } from "@/components/model/user/List";
+import { Condition } from "@/components/model/user/Condition";
 
 export const ListPage = () => {
   return (
     <BaseLayout>
-      <List />
+      <Container maxWidth="md">
+        <Paper>
+          <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
+            ユーザー一覧
+          </Typography>
+          <Divider />
+          {/* 検索条件 */}
+          <Condition />
+          <Divider />
+          {/* 検索結果 */}
+          <List />
+        </Paper>
+      </Container>
     </BaseLayout>
   );
 };
